+++ src/main/java/org/jabref/model/entry/Month.java
@@ -1,5 +1,9 @@
 package org.jabref.model.entry;
 
+import java.time.YearMonth;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+import java.util.Locale;
 import java.util.Optional;
 
 import org.jabref.model.strings.StringUtil;
@@ -84,7 +88,13 @@ public enum Month {
         if (testString.length() > 3) {
             testString = testString.substring(0, 3);
         }
+        
         Optional<Month> month = Month.getMonthByShortName(testString);
+        if (month.isPresent()) {
+            return month;
+        }
+
+        month = Month.parseGermanShortMonth(testString);
         if (month.isPresent()) {
             return month;
         }
@@ -93,6 +103,29 @@ public enum Month {
             int number = Integer.parseInt(value);
             return Month.getMonthByNumber(number);
         } catch (NumberFormatException e) {
+            return Optional.empty();
+        }
+    }
+
+    /**
+     * Parses a month having the string in German standard form such as
+     * "Oktober" or in German short form such as "Okt"
+     * 
+     * @param value,
+     *            a String that represents a month in German form
+     * @return the corresponding month instance, empty if input is not in German
+     *         form
+     */
+    private static Optional<Month> parseGermanShortMonth(String value) {
+        if ("Mae".equalsIgnoreCase(value) || "Maerz".equalsIgnoreCase(value)) {
+            return Month.getMonthByNumber(3);
+        }
+
+        try {
+            YearMonth yearMonth = YearMonth.parse("1969-" + value, 
+            		DateTimeFormatter.ofPattern("yyyy-MMM", Locale.GERMAN));
+            return Month.getMonthByNumber(yearMonth.getMonthValue());
+        } catch (DateTimeParseException e) {
             return Optional.empty();
         }
     }